requirements.txt
django==1.7
gunicorn
slumber
requests
responses
mock
djangorestframework
django-rest-swagger
django-filter
django-user-roles
django-jenkins
django-extensions

sphinx
sphinx_rtd_theme
